Port the source code to JS
