+++ .gitattributes
@@ -9,6 +9,7 @@
 *.json text
 *.md text diff=markdown
 *.mjs text
+*.ps1 text
 *.svg text
 *.ts text
 *.txt text
