+++ bin/which.ps1
@@ -1,0 +1,3 @@
+#!/usr/bin/env pwsh
+Set-StrictMode -Version Latest
+node "$PSScriptRoot/which.js" @args
