+++ jsconfig.json
@@ -1,14 +1,13 @@
 {
-	"include": ["*.js", "etc/*.js", "example/*.js", "src/**/*.ts", "test/**/*.js"],
+	"exclude": ["etc/eslint.config.js"],
+	"include": ["*.js", "bin/*.js", "etc/*.js", "example/*.js", "src/**/*.js", "test/**/*.js"],
 	"compilerOptions": {
-		"allowJs": true,
 		"baseUrl": ".",
 		"checkJs": true,
 		"esModuleInterop": true,
 		"isolatedModules": true,
 		"module": "NodeNext",
 		"moduleResolution": "NodeNext",
-		"noEmit": true,
 		"noImplicitOverride": true,
 		"resolveJsonModule": true,
 		"strict": true,
