+++ package.json
@@ -16,24 +16,23 @@
 		"which": "./bin/which.js"
 	},
 	"devDependencies": {
-		"@types/eslint__js": "^8.42.3",
+		"@babel/eslint-parser": "^7.24.1",
+		"@babel/plugin-syntax-import-attributes": "^7.24.1",
 		"@types/gulp": "^4.0.17",
 		"@types/node": "^20.12.7",
 		"del": "^7.1.0",
-		"esbuild": "^0.20.2",
 		"eslint": "^8.57.0",
 		"execa": "^8.0.1",
 		"gulp": "^5.0.0",
 		"typedoc": "^0.25.13",
-		"typescript": "^5.4.5",
-		"typescript-eslint": "^7.7.1"
+		"typescript": "^5.4.5"
 	},
 	"engines": {
-		"node": ">=20.0.0"
+		"node": ">=22.0.0"
 	},
 	"exports": {
 		"types": "./lib/index.d.ts",
-		"import": "./lib/index.js"
+		"import": "./src/index.js"
 	},
 	"files": [
 		"lib/",
@@ -48,7 +47,7 @@
 	],
 	"scripts": {
 		"prepack": "gulp",
-		"start": "gulp build && node bin/which.js --help",
-		"test": "gulp build && node --test --test-reporter=spec"
+		"start": "node bin/which.js --help",
+		"test": "node --test --test-reporter=spec"
 	}
 }
