+++ src/jsconfig.json
@@ -1,9 +1,10 @@
 {
-	"extends": "../tsconfig",
-	"include": ["index.ts"],
+	"extends": "../jsconfig",
+	"include": ["index.js"],
 	"compilerOptions": {
 		"declaration": true,
 		"declarationMap": true,
+		"emitDeclarationOnly": true,
 		"incremental": true,
 		"noEmit": false,
 		"outDir": "../lib",
