+++ src/result_set.js
@@ -8,30 +8,34 @@ export class ResultSet {
 
 	/**
 	 * The searched command.
+	 * @type {string}
+	 * @readonly
 	 */
-	readonly #command: string;
+	#command;
 
 	/**
 	 * The finder used to perform the search.
+	 * @type {Finder}
+	 * @readonly
 	 */
-	readonly #finder: Finder;
+	#finder;
 
 	/**
 	 * Creates a new result set.
-	 * @param command The searched command.
-	 * @param finder The finder used to perform the search.
+	 * @param {string} command The searched command.
+	 * @param {Finder} finder The finder used to perform the search.
 	 */
-	constructor(command: string, finder: Finder) {
+	constructor(command, finder) {
 		this.#command = command;
 		this.#finder = finder;
 	}
 
 	/**
 	 * Returns all instances of the searched command.
-	 * @returns All search results.
+	 * @returns {Promise<string[]>} All search results.
 	 */
-	async all(): Promise<string[]> {
-		const executables: string[] = [];
+	async all() {
+		/** @type {string[]} */ const executables = [];
 		for await (const path of this.stream()) if (!executables.includes(path)) executables.push(path);
 		if (executables.length) return executables;
 		throw Error(`No "${this.#command}" in (${this.#finder.paths.join(Finder.isWindows ? ";" : delimiter)}).`);
@@ -39,9 +43,9 @@ export class ResultSet {
 
 	/**
 	 * Returns the first instance of the searched command.
-	 * @returns The first search result.
+	 * @returns {Promise<string>} The first search result.
 	 */
-	async first(): Promise<string> {
+	async first() {
 		const {value} = await this.stream().next();
 		if (value) return value;
 		throw Error(`No "${this.#command}" in (${this.#finder.paths.join(Finder.isWindows ? ";" : delimiter)}).`);
@@ -49,9 +53,9 @@ export class ResultSet {
 
 	/**
 	 * Returns a stream of instances of the searched command.
-	 * @returns A stream of the search results.
+	 * @returns {AsyncGenerator<string, void>} A stream of the search results.
 	 */
-	stream(): AsyncGenerator<string, void> {
+	stream() {
 		return this.#finder.find(this.#command);
 	}
 }
